module-1/03_Logical_Branching/

+++ module-1/03_Logical_Branching/lecture/Lecture/04_ReturnDoubleOfTwo.cs
@@ -6,8 +6,9 @@
          4. This method currently returns an int. Change it so that it returns a double.
             TOPIC: Return Types
         */
-        public int ReturnDoubleOfTwo()
+        public double ReturnDoubleOfTwo()
         {
+          
             return 2;
         }
     }
